refactor: relocate home components
