+++ deno.lock
@@ -229,20 +229,19 @@
         "npm:@fontsource-variable/montserrat@^5.1.0",
         "npm:@fontsource-variable/saira@^5.1.0",
         "npm:@kobalte/core@~0.13.7",
-        "npm:@mdx-js/mdx@^3.1.0",
         "npm:@pandacss/dev@0.48",
         "npm:@solidjs/meta@~0.29.4",
         "npm:@solidjs/router@~0.15.1",
         "npm:@solidjs/start@^1.0.10",
         "npm:@types/dompurify@^3.2.0",
         "npm:@types/marked@6",
-        "npm:dompurify@^3.2.1",
+        "npm:dompurify@^3.2.2",
         "npm:husky@^9.1.7",
         "npm:lint-staged@^15.2.10",
-        "npm:marked@^15.0.2",
-        "npm:motion@^11.11.17",
+        "npm:marked@^15.0.3",
+        "npm:motion@^11.13.1",
         "npm:postcss@^8.4.49",
-        "npm:prettier@3.3.3",
+        "npm:prettier@3.4.1",
         "npm:rehype-autolink-headings@^7.1.0",
         "npm:rehype-sanitize@6",
         "npm:rehype-slug@6",
