+++ src/components/home/faq.tsx
