+++ src/components/home/how-to-start.tsx
